Feat: 에디터 제목 내용 분리
