+++ package.json
@@ -10,7 +10,25 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "@tiptap-pro/extension-table-of-content": "^2.3.0",
+    "@tiptap-pro/extension-unique-id": "^2.3.0",
+    "@tiptap/extension-blockquote": "^2.1.11",
+    "@tiptap/extension-bold": "^2.1.11",
+    "@tiptap/extension-bullet-list": "^2.1.11",
+    "@tiptap/extension-code-block": "^2.1.11",
     "@tiptap/extension-color": "^2.1.11",
+    "@tiptap/extension-document": "^2.1.11",
+    "@tiptap/extension-hard-break": "^2.1.11",
+    "@tiptap/extension-heading": "^2.1.11",
+    "@tiptap/extension-history": "^2.1.11",
+    "@tiptap/extension-horizontal-rule": "^2.1.11",
+    "@tiptap/extension-italic": "^2.1.11",
+    "@tiptap/extension-list-item": "^2.1.11",
+    "@tiptap/extension-ordered-list": "^2.1.11",
+    "@tiptap/extension-paragraph": "^2.1.11",
+    "@tiptap/extension-placeholder": "^2.1.11",
+    "@tiptap/extension-strike": "^2.1.11",
+    "@tiptap/extension-text": "^2.1.11",
     "@tiptap/extension-text-style": "^2.1.11",
     "@tiptap/extension-underline": "^2.1.11",
     "@tiptap/pm": "^2.1.11",
